feat(header.js): optimeze for laptop view
